medbot-rag/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pymongo==4.10.1
python-dotenv==1.0.1
tqdm==4.66.4
numpy==2.1.2
sentence-transformers==3.0.1
google-generativeai==0.7.2
pydantic==2.9.2
